Improved test.sh. Added new tests from lab-05

+++ lab_06_00/test.sh
@@ -1,8 +1,10 @@
 #!/bin/bash
 
-for i in {1..5}; do
-    ./lab.bin 'in_'$i'.txt' > last_test.txt 2>&1
-    diff -su 'out_'$i'.txt' last_test.txt
+for i in $(find . -name 'in_*.txt'); do
+    ./lab.bin $i > last_test.txt 2>&1
+    diff -su $(echo $i | sed 's/in/out/g') last_test.txt
 done
 
-rm last_test.txt
+if [ -f last_test.txt ]; then
+    rm last_test.txt
+fi
